capex-agent-demo/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pytest>=7.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
streamlit>=1.30.0
